LFO1 Envelope Lanes
The LFO1 Step Sequencer can retrigger the A and F envelope but
can't do so separately. This change allows it to do so.

- Separate the trigger_EG into _FEG and _AEG
- Handle multi-length trig state to trigger AEG or FEG separately
- A reasonable switching UI driven on shift-click to select only one
- Stream the full uint64

Addresses #1385. Not closing it yet since I want user feedback first.

+++ src/common/SurgeStorage.h
@@ -9,6 +9,7 @@
 #include <vector>
 #include <thread/CriticalSection.h>
 #include <memory>
+#include <stdint.h>
 
 #ifndef TIXML_USE_STL
 #define TIXML_USE_STL
@@ -386,7 +387,7 @@ struct StepSequencerStorage
    float steps[n_stepseqsteps];
    int loop_start, loop_end;
    float shuffle;
-   unsigned int trigmask;
+   uint64_t trigmask;
 };
 
 /*
